requirements.txt
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
python-multipart>=0.0.6,<1.0
websockets>=12.0,<13.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0
openai>=1.12.0,<2.0
anthropic>=0.18.0,<1.0
pytesseract>=0.3.10,<1.0
easyocr>=1.7.0,<2.0
Pillow>=10.2.0,<11.0
python3-discogs-client>=2.5.0,<3.0
musicbrainzngs>=0.7.1,<1.0
duckduckgo_search>=4.1.0,<5.0
trafilatura>=1.6.0,<2.0
lxml_html_clean>=0.4.0,<1.0
httpx>=0.26.0,<1.0
rapidfuzz>=3.5.0,<4.0
pyyaml>=6.0.1,<7.0
python-dotenv>=1.0.0,<2.0
structlog>=24.1.0,<25.0
opencv-python-headless>=4.9.0,<5.0
numpy>=1.26.0,<2.0
cachetools>=5.3.0,<6.0
beautifulsoup4>=4.12.0,<5.0
chromadb>=0.4.22,<1.0
PyMuPDF>=1.24.0,<2.0
tokenizers>=0.15.0,<1.0
